+++ compiler/src/org.graalvm.libgraal/src/org/graalvm/libgraal/LibGraalIsolate.java
@@ -40,15 +40,17 @@ import java.util.function.Supplier;
 public final class LibGraalIsolate {
 
     private final long id;
+    private final long address;
 
     private static final Map<Long, LibGraalIsolate> isolates = new HashMap<>();
 
-    static synchronized LibGraalIsolate forIsolateId(long isolateId) {
-        return isolates.computeIfAbsent(isolateId, a -> new LibGraalIsolate(a));
+    static synchronized LibGraalIsolate forIsolateId(long isolateId, long address) {
+        return isolates.computeIfAbsent(isolateId, id -> new LibGraalIsolate(id, address));
     }
 
-    private LibGraalIsolate(long isolateId) {
+    private LibGraalIsolate(long isolateId, long address) {
         this.id = isolateId;
+        this.address = address;
     }
 
     /**
@@ -101,7 +103,7 @@ public final class LibGraalIsolate {
         while ((cleaner = (Cleaner) cleanersQueue.poll()) != null) {
             cleaners.remove(cleaner);
             if (!cleaner.clean()) {
-                new Exception(String.format("Error releasing handle %d in isolate %d", cleaner.handle, id)).printStackTrace(System.out);
+                new Exception(String.format("Error releasing handle %d in isolate %d (0x%x)", cleaner.handle, id, address)).printStackTrace(System.out);
             }
         }
     }
@@ -128,6 +130,7 @@ public final class LibGraalIsolate {
     /**
      * Notifies that the {@code JavaVM} associated with {@code isolate} has been destroyed. All
      * subsequent accesses to objects in the isolate will throw an {@link IllegalArgumentException}.
+     * Called by {@code HotSpotGraalRuntime#shutdownLibGraal} using a JNI.
      */
     static synchronized void unregister(long isolateId) {
         LibGraalIsolate isolate = isolates.remove(isolateId);
@@ -138,7 +141,7 @@ public final class LibGraalIsolate {
 
     @Override
     public String toString() {
-        return String.format("%s[%d]", getClass().getSimpleName(), id);
+        return String.format("%s[%d (0x%x)]", getClass().getSimpleName(), id, address);
     }
 
     private boolean destroyed;